fix(scripts): dart and js releases
