+++ scripts/release/updateAPIVersions.ts
@@ -1,13 +1,13 @@
 import fsp from 'fs/promises';
 
-import yaml from 'js-yaml';
-
 import clientsConfig from '../../config/clients.config.json' assert { type: 'json' };
-import { CI, exists, GENERATORS, run, setVerbose, toAbsolutePath } from '../common.js';
+import { CI, exists, setVerbose, toAbsolutePath } from '../common.js';
 import { getGitHubUrl, getLanguageFolder } from '../config.js';
 import type { Language } from '../types.js';
 
 import { writeJsonFile } from './common.js';
+import { updateDartPackages } from './dart.js';
+import { updateJavaScriptPackages } from './javascript.js';
 import type { Changelog, Versions } from './types.js';
 
 async function updateConfigFiles(versionsToRelease: Versions): Promise<void> {
@@ -21,7 +21,7 @@ async function updateConfigFiles(versionsToRelease: Versions): Promise<void> {
   await writeJsonFile(toAbsolutePath('config/clients.config.json'), clientsConfig);
 }
 
-async function updateChangelog(
+export async function updateChangelog(
   lang: Language,
   changelog: string,
   current: string,
@@ -58,11 +58,9 @@ export async function updateAPIVersions(versions: Versions, changelog: Changelog
       continue;
     }
 
-    if (lang === 'javascript') {
+    if (lang === 'javascript' && releaseType) {
       setVerbose(CI);
-      await run(`yarn install && yarn release:bump ${releaseType}`, {
-        cwd: getLanguageFolder(lang),
-      });
+      await updateJavaScriptPackages(releaseType);
     }
 
     await updateChangelog(
@@ -72,86 +70,5 @@ export async function updateAPIVersions(versions: Versions, changelog: Changelog
       next,
       toAbsolutePath(`${getLanguageFolder(lang as Language)}/CHANGELOG.md`),
     );
-  }
-}
-
-/**
- * Updates packages versions and generates the changelog.
- */
-async function updateDartPackages(changelog: string, nextVersion: string): Promise<void> {
-  for (const gen of Object.values(GENERATORS)) {
-    if (gen.language !== 'dart') {
-      continue;
-    }
-
-    if (!nextVersion) {
-      throw new Error(`Failed to bump '${gen.packageName}'.`);
-    }
-
-    let currentVersion = await getPubspecField(gen.output, 'version');
-
-    // if there's no version then it mostly means it's a new client.
-    if (!currentVersion) {
-      currentVersion = '0.0.1';
-    }
-
-    await updateChangelog('dart', changelog, currentVersion, nextVersion, toAbsolutePath(`${gen.output}/CHANGELOG.md`));
-  }
-
-  // Version is sync'd on every clients so we set it once.
-  clientsConfig.dart.packageVersion = nextVersion;
-
-  // update `clients.config.json` file for the utils version.
-  await writeJsonFile(toAbsolutePath('config/clients.config.json'), clientsConfig);
-
-  // Core client package path
-  const corePackagePath = 'clients/algoliasearch-client-dart/packages/client_core';
-
-  // fetch the version from the pubspec file of the core package
-  let currentCoreVersion = await getPubspecField(corePackagePath, 'version');
-  if (!currentCoreVersion) {
-    currentCoreVersion = '0.0.1';
-  }
-
-  // update the changelog for core package
-  await updateChangelog(
-    'dart',
-    changelog,
-    currentCoreVersion,
-    nextVersion,
-    toAbsolutePath(`${corePackagePath}/CHANGELOG.md`),
-  );
-
-  // we've bumped generated clients but still need to do the manual ones.
-  await bumpPubspecVersion(toAbsolutePath(`${corePackagePath}/pubspec.yaml`), nextVersion);
-}
-
-/**
- * Get 'version' from pubspec.yaml file.
- */
-async function getPubspecField(filePath: string, field: string): Promise<string | undefined> {
-  try {
-    const fileContent = await fsp.readFile(toAbsolutePath(`${filePath}/pubspec.yaml`), 'utf8');
-    const pubspec = yaml.load(fileContent) as Record<string, any>;
-
-    return pubspec[field];
-  } catch (error) {
-    throw new Error(`Error reading file ${filePath}: ${error}`);
-  }
-}
-
-/**
- * Bump 'version' of the given pubspec.yaml file path.
- */
-async function bumpPubspecVersion(filePath: string, nextVersion: string): Promise<void> {
-  try {
-    const fileContent = await fsp.readFile(toAbsolutePath(filePath), 'utf8');
-    const pubspec = yaml.load(fileContent) as Record<string, any>;
-
-    pubspec.version = nextVersion;
-
-    await fsp.writeFile(filePath, yaml.dump(pubspec));
-  } catch (error) {
-    throw new Error(`Error writing file ${filePath}: ${error}`);
   }
 }
